Add more docs, move code
