+++ src/applications/vaos/covid-19-vaccine/redux/helpers/formSubmitTransformers.js
@@ -10,7 +10,7 @@ import {
   findCharacteristic,
   getClinicId,
   getSiteCode,
-} from '../../../services/healthcare-service/transformers';
+} from '../../../services/healthcare-service';
 
 export function transformFormToAppointment(state) {
   const clinic = getChosenClinicInfo(state);
