+++ src/applications/vaos/services/healthcare-service/transformers.js
@@ -149,45 +149,3 @@ export function transformAvailableClinics(facilityId, typeOfCareId, clinics) {
     transformAvailableClinic(facilityId, typeOfCareId, clinic),
   );
 }
-
-/**
- * Method to find a particular characteristic of a clinic.
- *
- * @param {HealthCareService} clinic
- * @param {string} characteristic The characteristic to search for.
- *
- * @returns {string} The display name of the characteristic search for or an empty string.
- */
-export function findCharacteristic(clinic, characteristic) {
-  const result = clinic?.characteristic.find(element => {
-    return element.text === characteristic;
-  });
-
-  return result?.coding?.code || result?.coding?.display;
-}
-
-function getIdentifierToken(clinic, index) {
-  // The FHIR clinic identifier format is, 'urn:va:healthcareservice:983:983:308'
-  // where the last 3 tokens represent the site id, facility id, and clinic id.
-  return clinic?.identifier[0].value.split(':')[index];
-}
-
-/**
- * Method to get the clinic id.
- *
- * @param {HealthCareService} clinic
- * @returns {string} The clinic id or empty string.
- */
-export function getClinicId(clinic) {
-  return getIdentifierToken(clinic, 5);
-}
-
-/**
- * Method to get the site code.
- *
- * @param {Object} clinic
- * @returns {String} The clinic site code or empty string.
- */
-export function getSiteCode(clinic) {
-  return getIdentifierToken(clinic, 3);
-}
